+++ core/src/telemetry/mod.rs
@@ -20,6 +20,9 @@ pub enum MetricCounter {
 	IncomingPutRecord,
 	IncomingGetRecord,
 	EventLoopEvent,
+	DHTPutRecords,
+	RpcConnected,
+	RpcConnectionSwitched,
 }
 
 pub trait MetricName {
@@ -40,6 +43,9 @@ impl MetricName for MetricCounter {
 			IncomingPutRecord => "light.incoming_put_record",
 			IncomingGetRecord => "light.incoming_get_record",
 			EventLoopEvent => "light.event_loop_event",
+			DHTPutRecords => "light.dht.put.records",
+			RpcConnected => "light.rpc.connected",
+			RpcConnectionSwitched => "light.rpc.connection_switched",
 		}
 	}
 }
@@ -57,6 +63,7 @@ impl MetricCounter {
 			MetricCounter::IncomingPutRecord,
 			MetricCounter::IncomingGetRecord,
 			MetricCounter::EventLoopEvent,
+			MetricCounter::DHTPutRecords,
 		]
 	}
 
@@ -65,6 +72,7 @@ impl MetricCounter {
 			(Origin::External, MetricCounter::Starts) => true,
 			(Origin::External, MetricCounter::Up) => true,
 			(Origin::External, _) => false,
+			(Origin::Internal, MetricCounter::DHTPutRecords) => false,
 			(_, _) => true,
 		}
 	}