+++ fat/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.12.7
+
+- Counting initialized and switched RPC host connections
+- Measure count of records put
+- Update `avail-light-core` to 1.2.7
+- Changed the exit codes on failure
+
 ## [1.12.6](https://github.com/availproject/avail-light/releases/tag/avail-light-fat-v1.12.6) - 2025-04-09
 
 - Update `avail-light-core` to 1.2.6